datamanagementsystem/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
djangorestframework_simplejwt>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
drf-yasg==1.21.7